Sora no Kiseki FC GOG installer x 2.
This commit is the next in a commit chain adding working installers for
the GOG version of the first Kiseki title, "Sora no Kiseki FC," patched
with Japanese voices converted from the "Evolution" Vita port via the
"SoraVoice" library. Specifically, this commit finalizes a first
(untested and thus probably non-working) draft for this installer.