Sora no Kiseki FC GOG installer x 4.
This commit is the next in a commit chain adding working installers for
the GOG version of the first Kiseki title, "Sora no Kiseki FC," patched
with Japanese voices converted from the "Evolution" Vita port via the
"SoraVoice" library. Specifically, this commit corrects most (if not
all) of our possibly working draft for this installer, which has yet to
be extensively tested and is thus probably still non-working.